Seating system
•	Seating System  - Android Application  -
Seating System integrates company seats’ plan, and
employees’ and managers’ details.
The main objective of this system is to permit managers in the company to add an employee in the system, the system also permits them to view their employees’ seats by floor and view those who work in the same project as them. In addition, the system targets allowing admins to view all employees in the company and filter them by manager or by project, it also aims to allow them to add managers and projects in the system.